# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.APIError:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  handler.AssignRoleRequest:
    properties:
      user_id:
        minimum: 1
        type: integer
    required:
    - user_id
    type: object
...
  handler.CreateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - display_name
    - name
    type: object
...
  handler.ListRolesResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      roles:
        items:
          $ref: '#/definitions/handler.RoleResponse'
        type: array
      total:
        type: integer
    type: object
...
  handler.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.UpdateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - display_name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nebula-live.com
    name: API Support
  description: Nebula Live is a modern backend API service built with Go, following
    Domain-Driven Design (DDD) principles and Clean Architecture.
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Nebula Live API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get list of roles with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            $ref: '#/definitions/handler.ListRolesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: List Roles
      tags:
      - RBAC Role Management
    post:
      consumes:
      - application/json
      description: Create a new role in the system
      parameters:
      - description: Role creation data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: Role already exists
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Create Role
      tags:
      - RBAC Role Management
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role from the system
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Role deleted successfully
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "403":
          description: Cannot delete system role
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Delete Role
      tags:
      - RBAC Role Management
    get:
      consumes:
      - application/json
      description: Get role information by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role retrieved successfully
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Get Role
      tags:
      - RBAC Role Management
    put:
      consumes:
      - application/json
      description: Update role information
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role update data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Update Role
      tags:
      - RBAC Role Management
  /roles/{id}/assign:
    post:
      consumes:
      - application/json
      description: Assign a role to a user
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: User assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/handler.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role assigned successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Role or user not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: Role already assigned
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Assign Role to User
      tags:
      - RBAC Role Management
  /roles/{id}/users/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a role from a user
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role removed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Role, user or user role not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Remove Role from User
      tags:
      - RBAC Role Management
  /roles/users/{userId}:
    get:
      consumes:
      - application/json
      description: Get all roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user roles
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.RoleResponse'
              type: array
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Get User Roles
      tags:
      - RBAC Role Management
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"